Add 15-inch MacBook Pro to Available Hardware.
Specs from https://www.apple.com/macbook-pro/specs.

I tried to choose as similar a configuration to the 13-inch model as possible,
given the model differences. In some places I chose "equal" rather than "max",
for example 16GiB of RAM, which is the max for the 13-inch even though the
15-inch allows up to 32GiB. I think this should be at or near to the lowest
cost for the 15-inch model.

Signed-off-by: M. J. Fromberger <michael.j.fromberger@gmail.com>