Go conventions: Use Go modules for applications
Update code convention to use Go modules for dependency management of applications. The change omits `vendor` intentionally. A separate discussion will be opened for this.

Also, highlight points for libraries/applications and update supported Go versions.

The reasoning is that we don't need an external tool like `dep` anymore to build applications. Our version policy for applications is to support one latest release of Go, which is 1.12, and it supports modules natively.

As noted in the [epic](https://github.com/src-d/backlog/issues/1392) (non-public link) most of our projects already switched to modules or are in the transition process. This change merely reflects the decision made previously and was already put into motion.