Use gender-neutral language and idiomatic phrasing.
Replace "he" with "they" where appropriate. Update to idiomatic English phrases
for business terms (e.g., "business days" instead of "labor days", "application
form" instead of "form to apply", and so forth).

There may still be some cases that could be further updated.  I was
intentionally conservative in my edits to avoid changing the meaning, and only
updated the text where it was clearly not an EU or Spanish legal term of art.

Otherwise, there are a few cosmetic changes (punctuation, whitespace) that were
incidental to the changes described above.

Signed-off-by: M. J. Fromberger <michael.j.fromberger@gmail.com>